:recycle: Fixes #875 -- Replace own Leaflet binding with react-leaflet
This library has far more extensive support for the entire life-cycle
and using it definitely solves the flicker for now.

Also painstakingly found out that in the Formio.js component destroy
method, we must clean up our ReactDOM nodes as otherwise we're leaking
memory and operating on map instances that aren't even part of the DOM
anymore.