:sparkles: [open-formulieren/open-forms#1155] -- Detect session expiry and display message
The components used between form start and form submission confirmation
page are wrapped in a session-expiry detection component which displays
an error message if the session has expired.

Note that the actual backend expiration is done in the backend through
the expiry of the session cookie, and we rely on meta-information passed
from backend API call responses to determine _when_ the session is about
to expire.