:poop: [open-formulieren/open-forms#1174] -- workaround for changing appointment data
We can deconstruct the backend data back into something that the select
understands, based on the passed in options.

Additionally, when changing an appointment, the original time is no
longer available and the smallest atomic thing that can change for
the appointment, so we clear that value to not display ISO-8601 select
options that are auto-created by Formio because it's not present in
the available options list.

All of this is a nasty hack on top of a nasty hack. We really have to
revisit how the appointments stuff works.