:sparkles: [open-formulieren/open-forms#807] -- include the document CSP nonce when making API calls
The API needs to respond with WYSIWYG generated HTML configured by
form designers, which may contain inline styles. The Open Forms backend
can post-process these into a nonced-style tag rather than inline
style attributes, but for this it needs to know the CSP nonce of the
parent document.

The party embedding the SDK is responsible for configuring their CSP
and passing the nonce to the SDK constructor.