:bug: [open-formulieren/open-forms#1274] -- fixes more event/race conditions
As suspected already, the formData ref can be refactored out, since the
formRef (if set) contains the same information (the submission data).

The updating of the formData ref lagged behind one tick to what the
actual submission data value was. This _can_ cause race conditions in
the logic check handling where values are set based on formData ref.

The surface for race conditions is now further reduced by relying on
the Formio instance in all cases.