:poop: Fixes open-formulieren/open-forms#1374 -- workaround protected evaluator
The protected evaluator has a different behaviour from eval-based
templates where coalesce-behaviour could be expressed in the template
itself.

This is a hacky solution where we pass a function for url template
rather than an actual template string, which is evaluated with the
context that the template is normally rendered with. Formio itself
does NOT expect a function to be used for the component.data.url
property, but the interpolation engine does support it. We have to
monkeypatch the startsWith function on the function object for Formio
to not crash, so this is a dirty solution.

Using different template expressions did not seem to work, and
pre-compiled templates are not an option either - they are functions
and they also depend on the dynamic other-field properties to configure
the query-lookups which are not known at build-time.