:bug: Fixes open-formulieren/open-forms#1494 -- correctly restore original canSubmit state on aborted logic checks
Given that:

1. a form was in a submittable state
2. user input in the form step is done in quick succession but the net
   end result is no changes (e.g. radio yes/no toggle to the other
   option and quickly back to the originally selected option)

Then the change handler caused the form submission to be blocked as
a safeguard against bypassing logic checks.

The logic check in turn shortcutted because the data being compared
was not different than the originally checked data, so no changes in
state are suspected.

However, this shortcut/guard clause failed to restore the original
canSubmit state of the form step, which was set to false by the
change handler.

This patch ensures that on early-exits of logic checks, the state is
correctly restored. Additionally, the canSubmit state for aborted
logic checks now no longer only depends on form validity, but also
on the previous state. Even though a client-side form is valid,
it may still be blocked from submission by logic rules.