:adhesive_bandage: Fixes open-formulieren/open-forms#1738 -- fix cause of dissapearing validation errors
A form component must be set to pristine=false for it to display errors.
The component is only marked as not-pristine after change events have
been processed by it.

Formio normally includes a submit button in its configuration, but our
custom submission handler/logic does not lead to the same
events/handlers being executed for the submission.

Basically, the whole form submission should be re-visited properly, but
that's challenging with the entire async validation flow. This bandaid
at least fixes the reported problem.