:pencil: [open-formulieren/open-forms#2253] -- update component play function
removed the 'double click' detection since we can't update the (global)
locale easily for the component to properly receive the 'current'
locale. This would rather have to be an explicit test case.

Updated the component to track state that an update is being processed,
and adapted the storybook tests for that.

This patch also breaks out the functional/smart component into its own
IntlProvider to control the isolation of the component properly.