:recycle: [open-formulieren/open-forms#2276] -- adopt utrecht form-field-description component
Replaced the custom openforms-errors styles with the utrecht-form-field-description--invalid
flavour and added open-forms specific overrides via an additional modifier.

The openforms-message class names were not used, so we could revert back to the vanilla
formio template for the message component. Quick check showed that styles do not seem
to be affected.