:bug: [open-formulieren/open-forms#2377] -- ensure link-styles are applied in WYSIWYG bodies
The WYSIWYG variant/modifier of our Body component did not correctly
apply link hover styles. This is now applied by styling all anchor tags
inside a WYSIWYG body as an utrecht-link component.

Note that the sass for this component is refactored a bit - the variants
and associated BEM modifiers are broken out into the component and
unused modifiers have been removed.

The WYSIWYG case has been added to the storybook documentation and the
proptypes are updated to correctly reflect the available modifiers.