:recycle: [open-formulieren/open-forms#2377] -- refactor the anchor variant of buttons
In total there are three anchor variants that need to be supported:

* without component specified (rendering an UtrechtButton)
* with anchor component specified (rendering an UtrechtButtonLink)
* with button component specified (rendering an UtrechtLinkButton)

The visual appearance of these is now set up so that they're all equal,
which we may have to revisit at some point. The button-theme mixin is
still used, as it addresses the various hover/focus/active states which
should be consistent with the other button variants and currently
override the upstream community styles in certain places.

The storybook button stories are extended with these variants so that
we have visual regression testing for the future.