:ambulance: Fixes open-formulieren/security-issues#19 -- escape textarea content to prevent self XSS
Self XSS was possible under a number of conditions:

* Textarea component used where XSS attack is entered/pasted
* The data is saved and then the user navigates back to the step
  to edit/view the data again
* Another component is present with a longer initialization sequence (like
  a date component with a calendar widget)

The third item is crucial as it causes formio to rebuild the entire webform _twice_
rather than once, which results in the now-updated data being available in the
context for the template, which happily renders the malicious payload.

In the other cases, the ready event fires quickly enough to only do one build
of the form, so the data never has a chance to end up in the templated out
code, while the value _is_ set via the DOM API and available in the textarea.

This patch addresses the edge case and ensures that the data is escaped
before passing it to the context.

Note that this is essentially a security issue in formio and has been
reported to them.

Backport-Of: 70340b3a626b2300816d9b2ad0369a29ba23d454