:iphone: [open-formulieren/open-forms#2676] -- Prevent overflow in long unbreakable words
This affects the progress indicator step titles, but also the title
of the step itself. The progress indicator effect is not limited to
mobile viewports only.

Additionally, the indicator marker items are no longer allowed to shrink
in their flexbox, keeping the icons + titles vertically aligned as they
have a fixed width set.