:clown_face: [#368] -- organize API mocks for stories and tests
This restructuring allows centralizing the API call mocks so
they can be used in (unit) tests AND in storybook stories without
needing to duplicate them.

When the API eventually changes, this gives us a single location
for the mocks to be updated.

Note that the component-specific mocks are placed close to the
component if possible, and some broader/tightly coupled
concepts like forms/submissions are placed in the generic
api-mocks directory.