:recycle: [open-formulieren/open-forms#3067] -- convert router root to data router
Date routers are the future of react-router. For now, we're gradually
adopting them and the way to do that is work our way down from the
root to the leaf nodes.

This change is required for the appointment routing with proper
relative URLs which uses hooks that are only available with data
routers.