:arrow_up: [open-formulieren/open-forms#3065] -- bump testing-library dependencies
With the upgrade to React 18, we can upgrade the testing-library tools
to their latest versions so that we can use the up-to-date documentation
and functionality.

userEvent.click was not triggering a number of things and this seems
to fix it.

Note that @storybook/testing-library pins it own dependencies to a
separate, older set and appears to still be working (at least in my
local env).