:alien: Convert .stories.mdx into CSF
Storybook has deprecated stories in .stories.mdx files, in favour of
defining them in Component Story Format (v3) and importing them
into MDX files.

This opportunity was taken to restructure the documentation a bit as
well.

The proze documentation now imports the relevant stories and is
a bit more summarized.