🎨[#3209] Use async syntax for clarity
Not really part of the ticket, but I based a spike on this test case to
see if a jest test behaves differently from a playbook interaction with
regards to `userEvent.hover()`. So at least we could have a jest
snapshot of the desired behaviour.

It does behave differently, but not in a useful way: in a jest test
after hover, there will be multiple tooltip elements added to the label
element, none of which is added to the document and therefore s all fail
`expect().toBeVisible()`.