:sparkles: [#435] -- implement component-type empty value
Formik state requires initial values for all (registered) fields so
that the components don't throw warnings because they're switching
from uncontrolled to controlled fields. For this, we need to be able
to retrieve the empty value for each supported component type.

Keep in mind these data types do not semantically match with the
data type of the component type, but have to be supported by (React)
DOM elements, e.g. '<input value={null} />' results in warnings,
it must be a string.