:sparkles: [open-formulieren/open-forms#3067] -- manage backend validation errors in container state
Any validation errors from the backend get stored in the parent component state,
which in turns makes them available (per step) for the relevant steps again to
set them as Formik initialErrors.

This exposes the handler to clear the backend errors from the state again when
the step is submitted (if you're just navigating between them, they are
not cleared). This is different from Formik's own client-side validation which
clears errors as part of a new validation cycle, triggered on blur of a field.