:recycle: Refactor useGetOrCreateSubmission hook
In the context of open-formulieren/open-forms#3067 - confirmation page.

The hook is responsible for creating a submission if none exists yet,
and it returned a 'loading: true' state when the check whether it
existed or not was being performed. This loading state ultimately
was not correct, as the async callback could result in 'nothing
to do' as outcome, but for a brief moment the state was updated
to loading.

We only want to reflect the loading state if we are actually doing
network IO, i.e. creating the submission in the backend, as an
alternative UI (spinner) is rendered when the loading state is
active. This causes the other (normal UI) component to mount/
unmount and can lead to UI flickering because it's fetching
its own data.