:bug: Prevent a child component remount
The RequireSession component was causing child component remounts when
that (via a hook) calls the session reset to reset the expiry date.

In particular, the  does network requests, which again
updates the session expiry from backend errors (TODO: fix this
in the backend to not emit the header) when it is mounted,
which leads to infinite loading loops.