:children_crossing: [#514] Always allow formik submission and validate on submit
Due to situations (hard refresh) where valid data is filled out, but
no changes are made, the form is not considered to be dirty, which
blocked the submit button. Relying on the valid state is not an
alternative option either, as this requires validation to run on
mount, which in turn discards the initialErrors set via the backend.

Additionally, running validation on blur/on change has some drawbacks
in the UI:
- browser autocomplete does not fire onBlur, which requires another
  click or [Tab] by the user to blur the active input for validadtion
  to trigger
- blur on the datepicker input triggers validation to early when the
  user is still selecting a date in the calendar (possibly navigating
  to the next/previous month)
- blur on the date input group may trigger validation too soon when
  a date part like day/month is being entered and no full date is
  yet available to validate
- running validation on change yields intrusive validation errors for
  email address input while the user is typing

Deferring the client-side validation to the actual submit action is
the easiest solution in this case.

Possible future improvement: scroll the first validation error into
view.