:recycle: [open-formulieren/open-forms#3203] Make SummaryConfirmation component responsible for getting config
This removes the duplicated code to retrieve the privacy policy
configuration from the backend, bringing it under the responsibility
of the SubmissionConfirmation component.

In the future, fetching this data should be moved to the data routers
loader, but since we have a mix of old and new style routers,
that's not possible yet.