:recycle: [open-formulieren/open-forms#3322] Calculate client-side entrypoint URL
During submission creation, the base URL of the form in the publicly
accessible frontend is recorded. This is used to be able to redirect
back from the backend to the frontend, wherever that may be hosted.

Using window.location.href introduces too much variability, as
appointments redirects to more appropriate sub-URLs and only then
the creation is triggered. We do not want to parse all these exceptions
in the backend.