:bug: [open-formulieren/open-forms#3404] Pass correct amount of products to backend
For each product being ordered, the amount of the product needs to be
passed along to the backend so that the appointment plugin can be
correctly informed of the amount of required time slots.

JCC actually limits the available time slots based on the amount of a
product being ordered.