:lipstick: Fix width of progress indicator
The bar was not taking up full width due to the negative margin-left
and margin-right values, combined with the width: 100% of the
parent. Setting the width to calc(100% + 2 * margin) results
in a horizontal scrollbar, but unsetting the width fixes the
styling on mobile while not destroying it on desktop.

Backport-of: #550