:recycle: [#571] Refactor away openforms-form-control
This wrapper was applied to the formio component node rendered by the
vanilla formio renderer, and copied over (for vertical spacing
reasons) to the React implementation of form field components.

The UtrechtFormField and UtrechtFormFieldset components already act
like a wrapper/container, and we've introduced a separate class name to manage
the vertical spacing of components. For interop with formio, we do
some CSS/SASS extension gymnastics, which we'll be able to remove
once we have our own React-based formio renderer.