:green_heart: Fix tests broken by DST
The storybook interaction tests are unreliable w/r to human readable
formatting of times and localizing them appropriately. Most notably,
it's hard to mock the current date in storybook interaction tests,
and it's further complicated by chromatic running tests on machines
for which we cannot control the TZ information.

The test is moved to jest with a fixed/mocked time so that we avoid
DST issues in the future, additionally the TZ information is pinned
to Europe/Amsterdam in the test script in package.json.