:adhesive_bandage: [open-formulieren/open-forms#3536] Fix crash because of requiring unvalidated data to be valid
The selected product IDs are required as query string inputs, but
the submitted data/user input at this point is not validated yet.

Using Zod to normalize the data causes crashes when the user leaves
the amount field empty or enters an invalid value (like 0 or a negative
value). Relaxing the validation schema is not an option, because
that causes problems downstream.

Instead, we discard the amount information as it wasn't used up until
now yet anyway and only pass the list of product IDs that are selected,
filtering out empty-ish values which accounts for no-choice-made-yet
in the product dropdown.