🐛 [open-formulieren/open-forms#3362] Handle server side redirects
The backend is now agnostic to URL routing/hash based routing - it will always
redirect back to the URL that was used to start the form and provide action &
action params query arguments that specify the target for the frontend.

The frontend parses this action & the params and maps it to the intended client
side routes, effectively decoupling implementation details between backend and
frontend.

Note that this requires the SDK to operate correctly in two of the tree main
steps in this flow:

1. SDK must correctly derive the 'base URL' for the form, irrespective of
   whether hash based routing is used or not. Fragments should *not* be sent
   to the backend, since they are ignored anyway.
2. The backend uses the URL supplied from 1. and append the action/action
   params from the context of the backend action/validation that was
   performed.
3. The SDK must correctly interpret the action and its params and route to
   the appropriate part of the application.

TODO: using this pattern, we can probably refactor the _start=1 flow from the
backend too, this can likely be converted to _of_action=startSubmission.