:bug: [open-formulieren/open-forms#3577] Fix checkbox/radio buttons on webkit
The label in our Open Forms theme stretches to the left side of the
checkbox, where the padding overlays the checkbox itself. On WebKit,
this blocks the click event/pointer event to the checkbox itself
(Chrome/Firefox are okay with this).

Swapping the order of the elements in the DOM still visually
positions them correctly while now allowing the input the lay
on top of the div/paragraph/label because it comes later in
the document tree.