:bug: Fix Leaflet styles
The layout/container refactor (in the backend) broke the map styles
in an 'interesting' way. We were loading the admin CSS for the form
builder in the public frontend too, and that refactor PR realised this
should not be needed and removed it. This admin CSS included the
Leaflet CSS to be able to render the map in the form builder. It also
provided the necessary CSS for the SDK maps, because it turns out the
SDK never properly bundled the Leaflet CSS.

Cleaning up in the backend exposed this issue in the SDK, so now the
import is moved to the SDK styles bundle to make it work reliably. This
also means that the map component probably never worked properly with
embedded forms on third party domains rather than served directly by
our own backend.