:pencil: [#36] Improve progress indicator item stories
Updated stories that aren't very 'logical' w/r to being completed but
also not-applicable at the same time.

Added additional annotations/docstrings and set up autodocs to describe
the various scenarios a particular state would occur in.

This also fixes the 'muted' styles that would apply to navigable
items (which is incorrect).