:recycle: [#36] Refactor/parametrize card styles
Rather than having the progress indicator component inherit the card
component styles, it shares the implementation details with it but
uses its own component namespace in the design tokens.

Implemented in a backwards compatible manner to point to existing
design tokens, but you can specify more specific ones if you'd like
so.