:bug: [#607] Patch regex for phone number validation
The first character of the phone number may only be a plus sign or
a number (4-digit phone numbers can be called, so only allowing a
zero would be too strict). For grouping, dashes and spaces may be
used, but these characters are not valid for the first character.

Backport-of: #608