:fire: Delete backported test
The combination of formio + testing library user-event v12 is impossible.

Can't even find any more documentation for this version of user-event,
since we're up to v14 by now. The fix is validated in Storybook and
we trust it without further tests (the tests are still on main and
other stable branches).