:poop: [open-formulieren/open-forms#3607] Wrap up this feature
This is *not* the level of quality we stand for, but it is costing
unreasonable amounts of time and effort to try to get a grip on the
interaction tests and state synchronization, without any guarantees
or conviction that we'll actually manage to sort it out.

Perhaps if we can upgrade Storybook and get testing-library v14 with
async user events things might get better, but that itself is a
non-trivial undertaking.

We now make sure that on every React re-render (due to validating
state changes) we also make sure to check if we need to fire the
formio blur event to trigger our formio-level validation.