[open-formulieren/open-forms#2617] Changed needsPayment to be retrieved from submission
The way to show the payment step in the progress indicator and to start
the payment form was through the form. Since we want to have access to
the price rules as well, we had to have access to the submission. So,
the submission is now the one which tells us if a form needs payment.