[open-formulieren/open-forms#3835] Fixed progress indicator concerning non applicable steps
The rendered steps and their data used to be retrieved from the form. With
this PR we do that by accessing both the form (startpage) and the
submission data in order to update correctly the PI if a step is
applicable or not (a logic rule may change this behaviour).