:sparkles: [#645] Implement standalone route/component for expired session
The component is responsible for cleanup of local/session storage and
resetting the session expiry if not done so yet.

Because the component is mounted on its own route, it will cause all
other components that trigger the redirect to be unmounted and their
state will automatically be reset too, leaving only the persistent
storages like session/local storage to be cleaned up.