:sparkles: [#4542] Wire up a callback for the verify button click
Passing a callback from the React-layer to the formio layer allows us
to avoid having to handle the whole interaction/modal situation in
formio, and we can implement this in the React layer.

The callback will be set up so that the verification modal can be
opened and it passes the necessary information (such as the entered
email and the component key being verified).