:bug: [open-formulieren/open-forms#4643] Reset step submit button block
When the email verification modal is opened, we can reset the blocked
submit step, which was blocked because of backend validation errors.

Even if the user does not actually complete the verification, the
server side validation will block the submission again and this
cycle repeats until the email is actually properly verified.