:sparkles: [open-formulieren/open-forms#4320] Support optional code for cosign login URL
If a code query parameter is present in the route params, use it to
pre-populate the code parameter for the ?next param, which redirects
to the submission look-up page on the backend. This will allow the
backend to shortcut the form by auto-populating it and validating the
authentication state, requiring less user input.

This page also works without the parameter and follows the 'old' flow.
The login screens will be updated accordingly to redirect to this page.