:children_crossing: [#4320] Update stories and UX of completion view
The backend will gain the ability to include the cosign link/button,
which will be rendered as a primary button. To point out the
right call to actions, the 'return to website' link has been
made a secondary button/action.

Stories have been updated to reflect the default templates in the
backend.