:truck: Rename .js files to .jsx
Vite requires the filename to be .jsx if React components are used
in them. Not all files have react components in them, but they
might in the future, and this way we only have a single rename
commit that we can easily ignore from the git blame, keeping
commit history properly intact now and in the future.