:hammer: [#724] Swap out Jest for Vitest
Set up a vite config and vitest configuration to transform our code
and setup into something that compiles and passes the tests. Create
React App is dead and the official React docs recommend using a
framework like Next.js/Redux or a toolchain like Vite when building
a low level library.

We start with migrating the test suite to the Vite ecosystem before
touching the bundler itself. This way we get most of the hiccups out
of the way, while (hopefully) minimizing impact on actual production
artifacts.

Jest configuration has been removed from package.json. It's currently
still used under the hood by storybook, but that too will be migrated
in due time.

We're sticking with jsdom for the time being - Vitest can run tests in
a real browser, but it's still marked as experimental and we don't
appear to be running into the same issues for MSW as we did with jest,
which required a workaround by using a fixed jsdom environment.