:alien: [#724] Fix token imports (CJS -> ESM)
Having looked into various configurations and plugins to automatically
get the CommonJS transformed into ESM breaks either dev mode or
production builds of Storybook.

There are some external plugins for CommonJS transformations, but they
either looks sketchy or don't seem to work in obvious ways (in
particular @rollup/plugin-commonjs wasn't a success).

So I've opted for the simple solution - write a custom plugin that
just replaces the CJS export with an ESM export and it works
in both dev and prod modes without any additional fuss. In the
future, we will process other formats of these token exports,
for which the design-token-editor package needs to be upgraded.